@@ -5,12 +5,13 @@ author:
 homepage: https://github.com/PROGrand/flutter_qa
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.3.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   rect_getter: ^0.1.0
+  built_collection:
 
 dev_dependencies:
   flutter_test:
