@@ -1,18 +1,22 @@
-
-
 abstract class MatchingInplaceAction {
-	int matchingIndex;
+  final int matchingIndex;
+
+  MatchingInplaceAction(this.matchingIndex);
 }
 
 class AddConnection extends MatchingInplaceAction {
-	int queryIndex;
-	int answerIndex;
+  final int sourceIndex;
+  final int destinationIndex;
+
+  AddConnection(this.sourceIndex, this.destinationIndex, int matchingIndex) : super(matchingIndex);
 }
 
 class RemoveConnection extends MatchingInplaceAction {
-	int queryIndex;
-	int answerIndex;
+  final int sourceIndex;
+
+  RemoveConnection(this.sourceIndex, int matchingIndex) : super(matchingIndex);
 }
 
 class RemoveAllConnections extends MatchingInplaceAction {
+  RemoveAllConnections(int matchingIndex) : super(matchingIndex);
 }
